+++ test/MockHttp.Tests/Language/Flow/Response/ClientTimeoutWithNullTimeSpanSpec.cs
@@ -1,0 +1,11 @@
+﻿#nullable enable
+namespace MockHttp.Language.Flow.Response;
+
+public class ClientTimeoutWithNullTimeSpanSpec : ClientTimeoutSpec
+{
+    public ClientTimeoutWithNullTimeSpanSpec()
+        : base(null)
+    {
+    }
+}
+#nullable restore