+++ test/MockHttp.Tests/Language/Flow/Response/ClientTimeoutWithZeroTimeSpanSpec.cs
@@ -1,0 +1,11 @@
+﻿#nullable enable
+namespace MockHttp.Language.Flow.Response;
+
+public class ClientTimeoutWithZeroTimeSpanSpec : ClientTimeoutSpec
+{
+    public ClientTimeoutWithZeroTimeSpanSpec()
+        : base(TimeSpan.Zero)
+    {
+    }
+}
+#nullable restore